# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: BTP User Containment API
  version: 0.1.0-SNAPSHOT
  description: |
    REST API for the OOBS: a central operator console
    that enrolls SAP BTP subaccounts / IAS tenants, discovers identity providers
    and SoD conflicts, and contains (locks) or unlocks compromised users across subaccounts,
    Cloud Foundry (XSUAA) and IAS in response to an incident.

    ## Authentication
    All `/api/v1/**` endpoints (except `/health` and `/whoami`) require a valid
    XSUAA-issued OAuth2 **bearer JWT**. Authorization is enforced per-operation
    via OAuth2 scopes; the required scope is noted in each operation description.

    ## CSRF
    Keep in mind that the SAP app router uses CSRF. 
    For that see: x-csrf-token docu on the sap help site 
    (it changes all x days so you need to search it)

    ## Scope model
    Four scopes exist, in increasing privilege. The `@PreAuthorize` rules map to:
    - **`btpc.viewer`** - read-only. Satisfied by any of the four scopes.
    - **`btpc.sod_viewer`** - may run SoD scans (admin also allowed).
    - **`btpc.responder`** - may contain/unlock users (admin also allowed).
    - **`btpc.admin`** - full control: enrollment, credentials, config, deletes.

    "Required scope" below names the *minimum* intent; `btpc.admin` always
    satisfies it.

    ## Caller source
    Mutating endpoints accept an optional `X-Caller-Source` header used only for
...
    not change behaviour or authorization.
  license:
...
  contact:
...
servers:
...
security:
  - xsuaa: []
paths:
...
  /api/v1/discovery/keys:
    get:
      tags: [Discovery]
      summary: List central-viewer keys
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Central-viewer keys
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CentralViewerKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Discovery]
      summary: Save a central-viewer key
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CentralKeySaveRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CentralViewerKey' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/discovery/keys/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Discovery]
      summary: Get one central-viewer key
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Central-viewer key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CentralViewerKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Discovery]
      summary: Update central-viewer key settings
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CentralKeyUpdateRequest' }
      responses:
        '200':
          description: Updated key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CentralViewerKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Discovery]
      summary: Delete a central-viewer key
      description: 'Required scope: `btpc.admin`.'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/discovery/keys/{id}/sync:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Discovery]
      summary: Trigger a discovery sync for one key
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      responses:
        '200':
          description: Sync result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/BadGateway' }
  /api/v1/discovery/candidates:
    get:
      tags: [Discovery]
      summary: List discovered subaccount candidates
      description: 'Required scope: `btpc.viewer`.'
      parameters:
        - name: onlyPromotable
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: centralKeyId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Candidates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DiscoveredSubaccount' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/discovery/candidates/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Discovery]
      summary: Get one discovery candidate
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Candidate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscoveredSubaccount' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Discovery]
      summary: Delete a discovery candidate
      description: 'Required scope: `btpc.admin`.'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------- CIS
...
components:
  securitySchemes:
    xsuaa:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        XSUAA-issued OAuth2 access token. Scopes (`btpc.admin`,
        `btpc.responder`, `btpc.sod_viewer`, `btpc.viewer`) are carried as
        `SCOPE_<name>` authorities and enforced per operation.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CallerSource:
      name: X-Caller-Source
      in: header
      required: false
      description: Audit attribution only; default UI.
      schema:
        type: string
        enum: [soar, soar-api, system]
  responses:
    BadRequest:
      description: Invalid input or malformed JSON body
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid bearer token
    Forbidden:
      description: Authenticated but missing the required scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    BadGateway:
      description: Upstream BTP/CIS failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Uniform error envelope. `null` fields are omitted.
      properties:
        error:
          type: string
          description: Stable machine code.
          enum: [bad_request, malformed_body, not_found, conflict, cis_upstream, forbidden, server_error]
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # ---- Platform ----
...
    CentralKeySaveRequest:
      type: object
      description: CentralKeyService.SaveRequest. The service key must classify as a CIS central-viewer key.
      required: [serviceKey]
      properties:
        serviceKey: { type: string, description: Raw CIS central-viewer service-key JSON (string). }
        label: { type: string, nullable: true }
        globalAccountName: { type: string, nullable: true }
        syncIntervalMinutes: { type: integer, default: 60 }
        syncEnabled: { type: boolean, default: true }
    CentralKeyUpdateRequest:
      type: object
      description: CentralKeyService.UpdateRequest. Partial - only sync settings and label.
      properties:
        syncEnabled: { type: boolean, nullable: true }
        syncIntervalMinutes: { type: integer, nullable: true }
        label: { type: string, nullable: true }
    CentralViewerKey:
      type: object
      description: ninja.gruber.btpc.discovery.domain.CentralViewerKey. The encrypted key material is never returned.
      properties:
        id: { type: string, format: uuid }
        globalAccountId: { type: string, format: uuid, nullable: true }
        globalAccountName: { type: string, nullable: true }
        label: { type: string, nullable: true }
        keyVersion: { type: integer }
        syncEnabled: { type: boolean }
        syncIntervalMinutes: { type: integer }
        lastSyncAt: { type: string, format: date-time, nullable: true }
        lastSyncError: { type: string, nullable: true }
        lastSyncCount: { type: integer, nullable: true }
        addedBy: { type: string }
        addedAt: { type: string, format: date-time }
    SyncResult:
      type: object
      description: DiscoveryService.SyncResult.
      properties:
        centralKeyId: { type: string, format: uuid }
        ok: { type: boolean }
        upserts: { type: integer, description: Number of candidate rows inserted/updated. }
        error: { type: string, nullable: true }
    DiscoveredSubaccount:
      type: object
      description: ninja.gruber.btpc.discovery.domain.DiscoveredSubaccount.
      properties:
        id: { type: string, format: uuid }
        centralKeyId: { type: string, format: uuid }
        subaccountGuid: { type: string, format: uuid }
        displayName: { type: string, nullable: true }
        subdomain: { type: string, nullable: true }
        region: { type: string, nullable: true }
        parentType: { type: string, nullable: true }
        parentGuid: { type: string, format: uuid, nullable: true }
        globalAccountGuid: { type: string, format: uuid, nullable: true }
        state: { type: string, nullable: true }
        stateMessage: { type: string, nullable: true }
        betaEnabled: { type: boolean, nullable: true }
        usedForProduction: { type: string, nullable: true }
        description: { type: string, nullable: true }
        discoveredAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        enrolledSubaccountId:
          type: string
          format: uuid
          nullable: true
          description: Set when this candidate is already enrolled; null means promotable.

    # ---- CIS ----
...